Added most of the logic, now to whack it into shape

+++ WebContent/js/init.js
@@ -207,29 +207,53 @@ function deleteFromCart(item) {
 
 }
 
+function updateShipping(type) {
+
+    $("#save-ship").hide();
+    $("#wc24-ship").hide();
+
+    if (type === "save")
+        $("#save-ship").show();
+    else
+        $("#wc24-ship").show();
+}
+
 function checkoutSave() {
-	
-	if ($("#filename").val() !== "") {
-		//Fill form with localStorage if available
-		
-		if (localStorage.getItem("game") !== null)
-			$("#cartg").val(localStorage.getItem("game"));
-		
-		if (localStorage.getItem("manga") !== null)
-			$("#cartm").val(localStorage.getItem("manga"));
-		
-		if (localStorage.getItem("record") !== null)
-			$("#cartr").val(localStorage.getItem("record"));
-		
-		
-		$("#checkout-form").submit();
-		
-		//Empty localStorage and submit the form.
-		localStorage.clear();
-		$("#checkout-confirm").html('<div class="progress"><div class="indeterminate amber"></div></div>Thanks for using DoujinSoft! Your save file will automatically download once done.');
-		
-	}
-	else
-		alert("Please upload a savefile before checking out your content.")
-	
+
+	//Fill form with localStorage if available
+    if (localStorage.getItem("game") !== null)
+        $("#cartg").val(localStorage.getItem("game"));
+
+    if (localStorage.getItem("manga") !== null)
+        $("#cartm").val(localStorage.getItem("manga"));
+
+    if (localStorage.getItem("record") !== null)
+        $("#cartr").val(localStorage.getItem("record"));
+
+
+    if ($("#save-radio")[0].checked) {
+
+        if ($("#filename").val() !== "") {
+
+            $("#checkout-form").submit();
+            localStorage.clear();
+            $("#checkout-confirm").html('<div class="progress"><div class="indeterminate amber"></div></div>Thanks for using DoujinSoft! Your save file will automatically download once done.');
+        }
+        else
+            M.toast({html:"Please upload a savefile before checking out your content."});
+    }
+
+    if ($("#wc24-radio")[0].checked) {
+
+        if ($("#recipient").val() !== "") {
+
+           $("#checkout-form").submit();
+           localStorage.clear();
+           $("#checkout-confirm").html('<div class="progress"><div class="indeterminate amber"></div></div>Thanks for using DoujinSoft! Look out for the blue light on your Wii.');
+        }
+        else
+           M.toast({html:"Please enter a Wii Friend Code before checking out your content."});
+
+    }
+
 }
